% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Impute_functions.R
\name{calculate_median}
\alias{calculate_median}
\title{Calculate Median for Each Column}
\usage{
calculate_median(data)
}
\arguments{
\item{data}{A data frame.}
}
\value{
A named vector of median values for each numeric column.
}
\description{
Calculates the median for each numeric column in a data frame, ignoring missing values.
}
