% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Detect Missing Value.R
\name{detect_missing_values}
\alias{detect_missing_values}
\title{Detect Missing Values}
\usage{
detect_missing_values(data)
}
\arguments{
\item{data}{A dataframe}
}
\value{
A named numeric vector with counts of missing values per column
}
\description{
Detect Missing Values
}
\examples{
detect_missing_values(airquality)
}
