% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Replave Missing with Median.R
\name{replace_missing_with_median}
\alias{replace_missing_with_median}
\title{Replace Missing Values with Column Medians}
\usage{
replace_missing_with_median(data)
}
\arguments{
\item{data}{A data frame with potentially missing values.}
}
\value{
A data frame with missing values replaced by medians.
}
\description{
Replaces missing values in each numeric column with that column's median.
}
