% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Impute_functions.R
\name{detect_missing_values}
\alias{detect_missing_values}
\title{Detect Missing Values in Each Column}
\usage{
detect_missing_values(data)
}
\arguments{
\item{data}{A data frame.}
}
\value{
A named vector with the count of missing values in each column with NAs.
}
\description{
Counts the missing values in each column of a data frame.
}
